old-tests/submission/Microsoft/video/video_019.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'ratechange' event fires when 'playbackRate' changes.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-defaultplaybackrate" />
  <meta name="assert" content="'ratechange' event fires when 'playbackRate' changes." />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script type="text/javascript">

  
  
  var t = async_test("'ratechange' event fires when 'playbackRate' changes.");
  
  function timeout()
  {
    t.step(function(){ assert_false(true, "'ratechange' event never fired!")});
    t.done();
  }
  
  var timer = setTimeout(timeout, 5000);
  
  function do_ratechange(event)
  {
        clearTimeout(timer);
        t.step(function(){ assert_equals(event.type, "ratechange")});
        t.done();
  }
  </script>
  <div id=log></div>
  <video id='video0' onratechange='do_ratechange(event);'>
  Your browser does not support media elements.
  </video>
  <script type="text/javascript">
    document.getElementById("video0").playbackRate = 2.0;
  </script>
</div>
</body>
</html>
